src/main/java/au/com/sealink/quicktravel/client/models/barcodes/ConsumerSplit.java
package au.com.sealink.quicktravel.client.models.barcodes;

import au.com.sealink.quicktravel.client.models.barcodes.core.Passenger;
import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

import java.util.ArrayList;
import java.util.List;

public class ConsumerSplit extends Base {
    @SerializedName("vehicles")
    @Expose
    private List<Object> vehicles = new ArrayList<Object>();

    @SerializedName("passengers")
    @Expose
    private List<Passenger> passengers = new ArrayList<Passenger>();

    public List<Object> getVehicles() {
        return vehicles;
    }

    public void setVehicles(List<Object> vehicles) {
        this.vehicles = vehicles;
    }

    public List<Passenger> getPassengers() {
        return passengers;
    }

    public void setPassengers(List<Passenger> passengers) {
        this.passengers = passengers;
    }
}
